+++ services/http/middleware/GetUserByToken.js
@@ -7,6 +7,11 @@ class GetUserByToken extends AbstractMiddleware {
 
   async middleware(req, res, next) {
     let { token } = req.body;
+    console.verbose(
+      `GetUserByToken token in BODY ${token}. Token if Authorization header ${req.get(
+        'Authorization',
+      )}`,
+    );
     if (!token) {
       token = req.get('Authorization');
       if (!token || token === 'null') {