+++ CHANGELOG.md
@@ -1,6 +1,5 @@
 ### 3.0.0
 
-[UPDATE] update deps
 [BREAKING] Mongoose v6. Than a lot of changes:[mongoDB drive changes](https://github.com/mongodb/node-mongodb-native/blob/4.0/docs/CHANGES_4.0.0.md), [Mongoose changes](https://mongoosejs.com/docs/migrating_to_6.html).
 Notable changes from migration
 Removed `execPopulate()`[link](https://mongoosejs.com/docs/migrating_to_6.html#removed-execpopulate)
@@ -42,6 +41,9 @@ await doc.populate([
   message: 'Some error';
 }
 ```
+
+[UPDATE] update deps
+[UPDATE] winston console transport now using timestapms
 
 ### 2.18.0